move IN_BLOB and MASKBITS to bits.py
